+++ src/FrontDoor/FrontDoor.Test/ScenarioTests/WebApplicationFireWallPolicyTests.ps1
@@ -25,7 +25,7 @@ function Test-PolicyCrud
     $matchCondition1 = New-AzFrontDoorMatchConditionObject -MatchVariable RequestHeader -OperatorProperty Contains -Selector "UserAgent" -MatchValue "Windows"
     $customRule1 = New-AzFrontDoorCustomRuleObject -Name "Rule1" -RuleType MatchRule -MatchCondition $matchCondition1 -Action Block -Priority 2
 
-    $ruleOverride = New-AzFrontDoorManagedRuleOverrideObject -RuleId "942100" -Action Log -EnabledState Enabled
+    $ruleOverride = New-AzFrontDoorManagedRuleOverrideObject -RuleId "942100" -Action Log
     $override1 = New-AzFrontDoorRuleGroupOverrideObject -RuleGroupName SQLI -ManagedRuleOverride $ruleOverride
     $managedRule1 = New-AzFrontDoorManagedRuleObject -Type DefaultRuleSet -Version "preview-0.1" -RuleGroupOverride $override1
 
@@ -41,7 +41,7 @@ function Test-PolicyCrud
     Assert-AreEqual $managedRule1.RuleGroupOverrides[0].ManagedRuleOverrides[0].Action $retrievedPolicy.ManagedRules[0].RuleGroupOverrides[0].ManagedRuleOverrides[0].Action
 
     $customRule2 = New-AzFrontDoorCustomRuleObject -Name "Rule2" -RuleType MatchRule -MatchCondition $matchCondition1 -Action Log -Priority 2
-    $updatedPolicy = Set-AzFrontDoorFireWallPolicy -Name $Name -ResourceGroupName $resourceGroupName -Customrule $customRule2
+    $updatedPolicy = Update-AzFrontDoorFireWallPolicy -Name $Name -ResourceGroupName $resourceGroupName -Customrule $customRule2
     Assert-NotNull $updatedPolicy
     Assert-AreEqual $Name $updatedPolicy.Name
     Assert-AreEqual $customRule2.Name $updatedPolicy.CustomRules[0].Name
@@ -67,14 +67,14 @@ function Test-PolicyCrudWithPiping
     $matchCondition1 = New-AzFrontDoorMatchConditionObject -MatchVariable RequestHeader -OperatorProperty Contains -Selector "UserAgent" -MatchValue "Windows"
     $customRule1 = New-AzFrontDoorCustomRuleObject -Name "Rule1" -RuleType MatchRule -MatchCondition $matchCondition1 -Action Block -Priority 2
 
-    $ruleOverride = New-AzFrontDoorManagedRuleOverrideObject -RuleId "942100" -Action Log -EnabledState Enabled
+    $ruleOverride = New-AzFrontDoorManagedRuleOverrideObject -RuleId "942100" -Action Log
     $override1 = New-AzFrontDoorRuleGroupOverrideObject -RuleGroupName SQLI -ManagedRuleOverride $ruleOverride
     $managedRule1 = New-AzFrontDoorManagedRuleObject -Type DefaultRuleSet -Version "preview-0.1" -RuleGroupOverride $override1
 
     New-AzFrontDoorFireWallPolicy -Name $Name -ResourceGroupName $resourceGroupName -Customrule $customRule1 -ManagedRule $managedRule1 -EnabledState Enabled -Mode Prevention
 
     $customRule2 = New-AzFrontDoorCustomRuleObject -Name "Rule2" -RuleType MatchRule -MatchCondition $matchCondition1 -Action Log -Priority 2
-    $updatedPolicy = Get-AzFrontDoorFireWallPolicy -Name $Name -ResourceGroupName $resourceGroupName | Set-AzFrontDoorFireWallPolicy -Customrule $customRule2
+    $updatedPolicy = Get-AzFrontDoorFireWallPolicy -Name $Name -ResourceGroupName $resourceGroupName | Update-AzFrontDoorFireWallPolicy -Customrule $customRule2
     Assert-NotNull $updatedPolicy
     Assert-AreEqual $Name $updatedPolicy.Name
     Assert-AreEqual $customRule2.Name $updatedPolicy.CustomRules[0].Name